mu4e: Top level maildir allows mail updates for all accounts
If the maildir is a narrow one, the `mu4e-get-mail-command` updates
the mails for _the selected context_. Although offlineimap (or any
other program) fetches the mails, mu4e will not update mu's
index. When the maildir is the top level one, which has accounts as
subdirs, mu4e manages to update all the accounts.