HTTPS only for melpa
On #emacs parsnip said why do you have it? I realised he's right and I
need to remove it. I use Linux systems only, so I have no use for win
specific code here. Please note that I'm not removing it because it is
"bad practise". I don't care. That code is useless for me, so it goes away.